+++ tests/Rule/GuessableDataTest.php
@@ -94,7 +94,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData($wordConverter);
+        $rule = new GuessableData([$wordConverter]);
 
         self::assertFalse($rule->test(new Password('pine4ppl€jack', ['apple'])));
         self::assertTrue($rule->test(new Password('pine4pp1€jack', ['apple'])));
@@ -112,7 +112,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData($wordConverter);
+        $rule = new GuessableData([$wordConverter]);
 
         self::assertFalse($rule->test(new Password('foo2B/II/1Bbar', [new DateTime('2018-11-28')])));
         self::assertTrue($rule->test(new Password('fooZB/I!/1Bbar', [new DateTime('2018-11-28')])));
@@ -137,7 +137,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData($wordConverter1, $wordConverter2);
+        $rule = new GuessableData([$wordConverter1, $wordConverter2]);
 
         self::assertTrue($rule->test(new Password('pine4ppl€jack', ['apple'])));
         self::assertFalse($rule->test(new Password('pineappl€jack', ['apple'])));
@@ -163,7 +163,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData($wordConverter1, $wordConverter2);
+        $rule = new GuessableData([$wordConverter1, $wordConverter2]);
 
         self::assertTrue($rule->test(new Password('foo2B/I!/1Bbar', [new DateTime('2018-11-28')])));
         self::assertFalse($rule->test(new Password('foo28/II/18bar', [new DateTime('2018-11-28')])));